% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two values}
\usage{
add(a, b)
}
\arguments{
\item{a}{numeric value}

\item{b}{numeric value}
}
\value{

}
\description{
\code{add} returns the sum of it's two arguments.
}
\examples{
add(2,2)

add(-1,5)

}
