% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycode.R
\name{add_strings}
\alias{add_strings}
\title{add strings}
\usage{
add_strings(strings)
}
\arguments{
\item{strings}{a vector of strings to concatinate}
}
\value{
a single string
}
\description{
Add strings by concatinating them together
}
\examples{
add_strings(c("abc", "def"))

}
