% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aretes.R
\name{ajout_arete}
\alias{ajout_arete}
\title{ajout d'arêtes}
\usage{
ajout_arete(grid, type, i, j)
}
\arguments{
\item{grid}{matrice de jeu}

\item{type}{arête horizontale ou verticale}

\item{i}{coordonnée abscisse}

\item{j}{coordonnée ordonnée}
}
\value{
matrice des arêtes horizontales et matrice des arêtes verticales
}
\description{
ajout d'arêtes
}
\examples{
ajout_arete(generate_slitherlink(6,6),"h",2,3)

}
