% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{bingxueFunction}
\alias{bingxueFunction}
\title{bingxueFunction}
\usage{
bingxueFunction(x)
}
\arguments{
\item{x}{a numeric value}
}
\description{
This is a function that produces the square of a given number
}
\details{
Additional details...
}
\examples{
inst/examples/bingxueFuntiongHelp.R

}
\author{
Bingxue
}
