% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius conversion
Convert degrees Celsius temperatures to degrees Fahrenheit}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{numeric variable}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Celsius conversion
Convert degrees Celsius temperatures to degrees Fahrenheit
}
