% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticas.R
\name{calcular_estatisticas}
\alias{calcular_estatisticas}
\title{Calcular estatisticas do dataset gapminder}
\usage{
calcular_estatisticas(data, continentes, anos = 2000)
}
\arguments{
\item{data}{Dados}

\item{continentes}{Continentes a serem avaliados}

\item{anos}{Periodo a ser avaliado}
}
\value{
Estatisticas baseadas no continente e periodo escolhido
}
\description{
Calcular estatisticas baseadas no continente e periodo escolhido
}
