% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estatisticas.R
\name{calcular_estatisticas_gapminder}
\alias{calcular_estatisticas_gapminder}
\title{Calcular estatisticas do dataset gapminder}
\usage{
calcular_estatisticas_gapminder(continentes, anos = 2000)
}
\arguments{
\item{continentes}{Continentes a serem avaliados}

\item{anos}{Periodo a ser avaliado}
}
\value{
Estatisticas baseadas no continente e periodo escolhido
}
\description{
Calcular estatisticas baseadas no continente e periodo escolhido
}
\examples{
calcular_estatisticas_gapminder("Europa", 2000)
}
