% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_average_production_by_groups.R
\name{calculate_average_production_by_groups}
\alias{calculate_average_production_by_groups}
\title{Calculate the average production of a dataframe}
\usage{
calculate_average_production_by_groups(data, ...)
}
\arguments{
\item{data}{A dataframe with the column \code{production}.}

\item{...}{Variables to group by.}
}
\value{
A dataframe with the summarized column \code{average_production}.
}
\description{
This function computes the average production of a dataframe. Further
grouping variables can be passed to subsequent arguments.
}
\examples{
data <- tibble::tribble(
  ~sector,  ~technology, ~production,
  "power", "renewables",        100L,
  "power",      "hydro",         10L,
  "automotive",        "ice",          5L
)

calculate_average_production_by_groups(data)

calculate_average_production_by_groups(data, sector)

}
