% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noeuds.R
\name{check_noeuds}
\alias{check_noeuds}
\title{verif nooeuds}
\usage{
check_noeuds(g)
}
\arguments{
\item{g}{grille}
}
\value{
verifie qu'il n'y a pas de noeuds en s'assurant que chaque coin recoit soit 0 arête soit 2 arêtes
}
\description{
verif nooeuds
}
