% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{demo}
\alias{demo}
\alias{demo-package}
\title{R package for demo part in insight interview: demo}
\description{
This is the R package for demo part in insight interview
}
\details{
This is the R package for demo part in insight interview.
The package contain three functions and two datasets.
functions:
my_lm_int runs linear regression model for lifeExp w.r.t. year;
li_spline gets the linear spline of three variables w.r.t. year;
make_mov makes movie.
dataset:
gDat is the data of gdp, population and lifeExp of various countries;
abbr is the abbr of the name of the country.
}

