% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEMObuODlmTest.R
\name{dEMObuODlmTest}
\alias{dEMObuODlmTest}
\title{dEMO differential expression analysis from RNA-seq data sets.}
\usage{
dEMObuODlmTest(expr, condition, original = NULL, alpha = 0.05,
  method.adj = "BH")
}
\arguments{
\item{expr}{data.frame, ExpressionSet or matrix after dataset has been
filtered}

\item{condition}{Binary vector where 0 means control and 1 treatment}

\item{original}{data.frame, ExpressionSet or matrix before dataset has been
filtered}

\item{alpha}{significance level for hypothesis test. Default value is 0.05}

\item{method.adj}{It is the multiple testing correction method. Default value
correspond to Benjamini-Hochberg correction. c("holm", "hochberg",
"hommel", "bonferroni", "BH", "BY","fdr", "none")}
}
\value{
A data.frame is returned, which cointains log2FC, PVALUE, ADJ.PVAL
  (adjusted p value) and dEMO.STAT (statistic of dEMO test)
}
\description{
Differential Expression Analysis dEMO, which assums counts from
  RNA-seq experiments fit two different negative binomial distributions, as
  dEMO test method explains. Overdisperssion parameter is estimated by
  \code{\link[edgeR]{estimateTagwiseDisp.default}} function.
}
\examples{
library(compcodeR)
library(edgeR)
set.seed(123456987)
datasCI <- generateSyntheticData(dataset = dEMOresults, n.vars = 15000,
samples.per.cond = 6, n.diffexp = 1500,repl.id = 1, seqdepth = 1e7,
fraction.upregulated = 0.5, between.group.diffdisp = FALSE,
filter.threshold.total = 1, filter.threshold.mediancpm = 0,
fraction.non.overdispersed = 0)
expressiondata <- datasCI@count.matrix
TMMfac <- calcNormFactors.default(expressiondata, method = "TMM")
exprT <- t(t(expressiondata)*TMMfac)
conditions <- (datasCI@sample.annotations$condition - 1)
testdEMOTMM <- dEMObuODlmTest(expr = exprT, condition = conditions,
original = exprT)
}
\author{
Enrique Perez_Riesgo
}
