% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEMOdesign-package.R
\docType{package}
\name{DEMOdesign-package}
\alias{DEMOdesign}
\alias{DEMOdesign-package}
\title{DEMOdesign: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cyang728/DEMOdesign}
  \item Report bugs at \url{https://github.com/cyang728/DEMOdesign/issues}
}

}
\author{
\strong{Maintainer}: Cheng-Han Yang \email{chenghanyang728@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
