% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEMOOD1sci.R
\name{dEMOOD1sci}
\alias{dEMOOD1sci}
\title{Confidence interval from log2(p1/p2), feature by feature,}
\usage{
dEMOOD1sci(expr, condition, Fescd, n = NULL, cd, alpha = 0.05,
  method.adj = "BH")
}
\arguments{
\item{expr}{matrix where the first set of columns store the counts for each
sample and the second set of columns store the library size for each
sample. For instance, if "n1" is the library size from sample 1 and x1 is
the counts from sample 1, and j is the number of samples which belong to
the same condition that sample 1, the column 1 is x1 and "j+1" is n1, the
column 2 is x2 and "j+2" is n2, and so on.}

\item{condition}{Binary vector where 0 means control and 1 treatment}

\item{Fescd}{is the data.frame which stores coefficients calculated by Fesc
function from dEMO package}

\item{cd}{it is the common disperssion estimated by
\code{\link[edgeR]{estimateCommonDisp}} function from edgeR package}

\item{alpha}{significance level for hypothesis test. Default value is 0.05}

\item{method.adj}{It is the multiple testing correction method. Default value
correspond to Benjamini-Hochberg correction. c("holm", "hochberg",
"hommel", "bonferroni", "BH", "BY","fdr", "none")}
}
\value{
A data.frame is returned, which cointains log2FC, FC_LowBound and
  FC_UpBound
}
\description{
This function provides a confidencial interval from
  log2(proportions ratio) or log2(FC) where proportions belong to negative
  binomial population and from data sets by considering each row as a
  different experiment, from data sets by considering each row as a different
  experiment, where each row has nothing to do with the others rows.
  Overdispersion estimation is carried out with
  edgeR::estimateTagwiseDisp.default function. dEMO::EMOOD1sci function works
  row by row, so if you want to calculate dEMO confidence intervals when rows
  are independent experiments, a loop "for" is recomended, as is showed in
  the example
}
\details{
Sometimes, data sets store several rows where each of them does not
  came from the same experiment, that is, each row has nothing to do with the
  other rows. Conversely, the rows from RNA-seq data sets are dependent to
  the other rows into the same column (sample), so other function has to be
  apply here, such as \code{\link{EMObuODlmTest}}.
}
\examples{
#this example is the continuation of example from Fesc function
cd <- estimateCommonDisp(y = expressiondata[, 1:6])
for(z in 1:dim(expressiondata)[1]){
testdEMOODci <- dEMOOD1sci(expr = expressiondata[z, 1:6],
Fescd = Fescoefs, condition = c(0, 0, 0, 1, 1, 1),
n = expressiondata[z, 7:12], cd = cd, alpha = 0.05, method.adj = "none")
}
}
\seealso{
\code{\link{Fesc}} for use its outcomes as argument of this
  function and test the differential expression for independents rows.

Other One sample functions: \code{\link{Fesc}}
}
\author{
Enrique Perez_Riesgo
}
