% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEMOODci.R
\name{dEMOODci}
\alias{dEMOODci}
\title{Confidence interval from log2(p1/p2) or log2(Fold Change)}
\usage{
dEMOODci(expr, condition, original = NULL, alpha = 0.05,
  method.adj = "BH")
}
\arguments{
\item{expr}{data.frame, ExpressionSet or matrix after dataset has been
filtered}

\item{condition}{Binary vector where 0 means control and 1 treatment}

\item{original}{data.frame, ExpressionSet or matrix before dataset has been
filtered}

\item{alpha}{significance level for hypothesis test. Default value is 0.05}

\item{method.adj}{It is the multiple testing correction method. Default value
correspond to Benjamini-Hochberg correction. c("holm", "hochberg",
"hommel", "bonferroni", "BH", "BY","fdr", "none")}
}
\value{
A data.frame is returned, which cointains log2FC, PVALUE, ADJ.PVAL
  (adjusted p value) and dEMO.STAT (statistic of dEMO test)
}
\description{
This function provides a confidencial interval estimation from
  log2(proportions ratio) or log2(FC) where proportions belong to negative
  binomial population, such as RNA-seq data sets. Overdispersion estimation
  is carried out with \code{\link[edgeR]{estimateTagwiseDisp.default}} function.
}
\details{
function \code{\link{EMObuODlmTest}} is employed when data sets came
  from RNA-seq assays, where each row from expression matrix from RNA-seq
  data sets are dependent to the other rows into the same column (sample).
  Conversely, data sets store several rows where each of them does not came
  from the same experiment, that is, each row has nothing to do with the
  other rows, and \code{\link{dEMOOD1sci}} function has to be used.
}
\examples{
datasCI <- compcodeR::generateSyntheticData(dataset = dEMOresults,
n.vars = 15000,samples.per.cond = 6, n.diffexp = 1500,repl.id = 1,
seqdepth = 1e7,fraction.upregulated = 0.5,between.group.diffdisp = FALSE,
filter.threshold.total = 1,filter.threshold.mediancpm = 0,
fraction.non.overdispersed = 0)
expressiondata <- datasCI@count.matrix
TMMfac <- edgeR::calcNormFactors.default(expressiondata, method = "TMM")
exprT <- t(t(expressiondata)*TMMfac)
conditions <- datasCI@sample.annotations$condition
testdEMOODci <- dEMOODci(expr = exprT, condition = (conditions-1),original = exprT)
}
\author{
Enrique Perez_Riesgo
}
