% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dplot_rnorm}
\alias{dplot_rnorm}
\title{Random number generator and plotter}
\usage{
dplot_rnorm(n = 100)
}
\arguments{
\item{n}{number of random normal numbers you'd like generated (default is 100)}
}
\description{
This function generates random normal numbers, and creates a density plot for them.
}
\examples{
dplot_rnorm()
dplot_rnorm(n = 1000)
}
\keyword{normal,}
\keyword{plot}
\keyword{random}

