% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brasileirao.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Encontrar pior ano de um time}
\usage{
encontrar_pior_ano_time(time)
}
\arguments{
\item{time}{chr Nome de um time do campeonato brasileiro.}
}
\value{
tibble Uma tabela de uma linha.
}
\description{
Usando a base de partidas do Brasileirao,
encontra o pior ano de um time ate o ano de 2022
}
\examples{
encontrar_pior_ano_time("Santos")
encontrar_pior_ano_time("Sport")

}
