% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_temp.R
\name{F_to_C}
\alias{F_to_C}
\title{Package demo will be used for converting temperature from celsius to
Fahrenheit and vice versa
Fahrenheit conversion
Convert degrees Fahrenheit temperatures to degrees Celsius}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{numeric variable}
}
\value{
The temperature in degrees Celsius
}
\description{
Package demo will be used for converting temperature from celsius to
Fahrenheit and vice versa
Fahrenheit conversion
Convert degrees Fahrenheit temperatures to degrees Celsius
}
