% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FoToHfs.R
\name{FoToHfs}
\alias{FoToHfs}
\title{Function emulating data transfer from a front office system to Hadoop}
\usage{
FoToHfs(FO, Dateto, Fo.meta.data)
}
\arguments{
\item{Dateto}{Cut date for the pushed dataset}

\item{Fo.meta.data}{}

\item{character,}{supported values, "BoE" and "local";  Name of FO system}
}
\description{
This function emulates data transfer from a front office system to Hadoop file system.
}
\examples{
FoToHfs()
}

