% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fremontBikes.R
\docType{data}
\name{fremontBikes}
\alias{fremontBikes}
\title{Fremont Bike Counter}
\format{A data frame with XXX rows and XXX variables:
\describe{
  \item{date}{Date (character)}
  \item{fremont_bridge_total}{Date of data collection}
  \item{fremont_bridge_east_sidewalk}{Species of the tree.  All dead trees were listed as "unknown"}
  \item{fremont_bridge_west_sidewalk}{Common name of the tree}
}}
\source{
\url{https://www.portlandoregon.gov/parks/article/433143}
}
\usage{
fremontBikes
}
\description{
A dataset of XXX
}
\keyword{datasets}
