% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice.R
\name{generate_slitherlink}
\alias{generate_slitherlink}
\title{grille de jeu}
\usage{
generate_slitherlink(n, m, max_attempts = 1000)
}
\arguments{
\item{n}{nombre de lignes}

\item{m}{nombre de colonnes}

\item{max_attempts}{nombre max de boucles parcourues par la fonction pour essayer de créer une matrice correcte}
}
\value{
liste de la grille de jeu, la matrice du nombre d'arêtes horizontales et la matrice du nombre d'arêtes verticales
}
\description{
grille de jeu
}
\examples{
generate_slitherlink(6,6)
generate_slitherlink(4,4)
}
