% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateFoMetaData.R
\name{generateFoMetaData}
\alias{generateFoMetaData}
\title{Auxiliary function generating meta data for the FO system}
\usage{
generateFoMetaData()
}
\value{
List emulating FO meta data
}
\description{
Meta data for the FO system is expected to be generated prior to
data trasfer from FO to HFS
}
\examples{
generateFoMetaData()
}

