% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{hello}
\alias{hello}
\title{Hello function}
\usage{
hello(x = NULL)
}
\arguments{
\item{x}{A character specifying the mode of operation: 'ct' for central tendency,
'df' for data entry and presentation, 'mo' for mathematical operations.
Default is NULL.}
}
\description{
This function demonstrates different functionalities based on the input.
}
\examples{
hello("ct")
hello("df")
hello("mo")

}
