% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hfsToInMemory.R
\name{hfsToInMemory}
\alias{hfsToInMemory}
\title{Imports data from Hadoop file system to memory}
\usage{
hfsToInMemory(hfsFile)
}
\arguments{
\item{hfsFile}{file name in HFS}
}
\value{
A character vector of the names of objects created, invisibly.
}
\description{
Imports data from Hadoop file system to memory
}
\examples{
hfsToInMemory("Hadoop.Rdata")
}

