% Generated by roxygen2 (4.0.2): do not edit by hand
\name{li_spline}
\alias{li_spline}
\title{Linear Spline}
\usage{
li_spline(x, yr = 1952:2006)
}
\arguments{
\item{x,}{yr}
}
\value{
data frame of result of linear spline
}
\description{
This function gets the linear spline of three variables w.r.t. year.
}
\details{
This function gets the linear spline of three variables w.r.t. year.
Three default variables are pop, lifeExp and gdpPercap.
}
\note{
the data set must contain variable year, pop, lifeExp, gdpPercap
}
\examples{
li_spline(x)
}

