% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.url.BOE.R
\name{make.url.BOE}
\alias{make.url.BOE}
\title{URL for accessing Bank of England's data}
\usage{
make.url.BOE(SeriesCodes, Datefrom, Dateto, CSVF = "TN", UsingCodes = "Y",
  VPD = "Y", VFD = "N")
}
\arguments{
\item{SeriesCodes}{Comma separated list of full series codes (mandatory parameter), e.g. LPMAUZI,LPMAVAA}

\item{Datefrom}{DD/MON/YYYY (mandatory parameter), e.g. 01/Feb/2006}

\item{Dateto}{DD/MON/YYYY (mandatory parameter), e.g. 01/Feb/2007}

\item{CSVF}{"TT" (Tabular with titles), "TN" (Tabular no titles),
"CT" (Columnar with titles) or "CN" (Columnar no titles). (mandatory parameter)}

\item{UsingCodes}{"Y" (mandatory parameter)}

\item{VPD}{"Y" or "N" (optional parameter).
The "VPD=Y" parameter need only be included if provisional data is desired.}

\item{VFD}{"Y" or "N" (optional parameter).
       The "VFD=Y" parameter should be supplied if you want to see any observation
       footnotes which have been associated to the selected data observations.}
}
\value{
A string representing URL of the corresponding data on Bank of England's website,
       e.g. http://www.bankofengland.co.uk/boeapps/iadb/fromshowcolumns.asp?csv.x=yes&
       Datefrom=01/Feb/2006&Dateto=01/Oct/2007 &SeriesCodes=LPMAUZI,LPMAVAA&CSVF=TN&
       UsingCodes=Y&VPD=Y&VFD=N
}
\description{
Generates an URL for accessing Bank of England's data, and returns it as a string
}
\examples{
make.url.BOE("RPQTFHD","31/Mar/1993","06/Aug/2015")
}

