% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_mov}
\alias{make_mov}
\title{Make movie}
\usage{
make_mov(name, loc, pic, frames = 4)
}
\arguments{
\item{name,}{loc, pic, frames}
}
\value{
a video file
}
\description{
This function makes movie using existing pictures
in selected file
}
\details{
This function uses ffmpeg in the system and makes
movie by connecting pictures in selected file
}
\note{
names must end with an video file extension, like ".mp4". pic must be jpg file.
}
\examples{
draw_singleyear(1952)
}
\keyword{misc}

