% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_numeric_columns.R
\name{mean_of_numeric_columns}
\alias{mean_of_numeric_columns}
\title{Compute the mean of numeric columns in a data frame}
\usage{
mean_of_numeric_columns(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A data.frame.
}
\description{
Compute the mean of numeric columns in a data frame
}
\examples{
data <- tibble::tibble(x = 2, y = 4, z = 8)
mean_of_numeric_columns(data)
}
