% Generated by roxygen2 (4.0.2): do not edit by hand
\name{my_lm_int}
\alias{my_lm_int}
\title{Regression results}
\usage{
my_lm_int(data)
}
\arguments{
\item{data}{}
}
\value{
intercept, slope and max of residual
}
\description{
This function runs linear regression model for lifeExp w.r.t. year,
and returns the coefs and max of residual
}
\details{
This function runs linear regression model for lifeExp w.r.t. year,
and returns the coefs and max of residual. Three return values are intercept,
slope and max of residual, which is defined as residual of regression over
sigma of regression.
}
\note{
the data set must contain variable year and lifeExp
}
\examples{
my_lm_int(gDat)
}

