% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Data of Brazilian soccer matches}
\format{
A tibble with 8026 rows and 8 columns.
\describe{
\item{temporada}{chr season}
\item{data}{date date}
\item{time_casa}{chr home}
\item{gols}{chr goals}
\item{time_visitante}{chr away}
\item{gols_casa}{dbl goals home}
\item{gols_visitante}{dbl goals away}
\item{quem_ganhou}{chr winner}
}
}
\source{
\url{https://github.com/williamorim/brasileirao}
}
\usage{
partidas_brasileirao
}
\description{
partidas_brasileirao is a data set from Brazilian soccer matches.
}
\keyword{datasets}
