% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{plot.ls_optim}
\alias{plot.ls_optim}
\title{Plot fit from linear model}
\usage{
\method{plot}{ls_optim}(object, ...)
}
\arguments{
\item{object}{A \code{list} of class \code{ls_optim} with the results of the fit from the function \code{beta_hat}.}
}
\value{
A \code{plot} representing the line of best fit for all predictors.
}
\description{
This function creates plots of the line of best fit for each predictor
}
\author{
Roberto Molinari
}
