% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-with-dependency.R
\name{quakes_leaflet}
\alias{quakes_leaflet}
\title{Function with dependency}
\usage{
quakes_leaflet()
}
\description{
This function depends upon another package (leaflet), and you need to include it with devtools. For more information on leaflet, visit their website:
https://rstudio.github.io/leaflet/
}
\keyword{leaflet}

