% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{sim_linear}
\alias{sim_linear}
\title{Simulate data from a linear model}
\usage{
sim_linear(n, beta, sigma2, seed = NA)
}
\arguments{
\item{n}{A \code{double} value representing the desired sample size.}

\item{beta}{A \code{double} value of the vector containing the true parameter values of the model.}

\item{sigma2}{A \code{double} value representing the desired variance of the residuals of the model.}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{y}{The \code{double} vector of the \code{n} simulated responses from the linear model}
 \item{X}{The \eqn{n \times 4} \code{double} value of the matrix whose columns represent the different predictors}
}
}
\description{
This function simulates data of a response coming from a linear model
}
\author{
Roberto Molinari
}
