% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Read R code from all files in a directory}
\usage{
sourceDir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{Full path to the directory}

\item{trace}{logical: if TRUE, each file name is printed before sourcing}

\item{...}{}
}
\value{
TBC
}
\description{
Sources all R codes from from all .[RrSsQq] files in a directory
}
\examples{
sourceDir(".")
}

