% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vennplot}
\alias{vennplot}
\title{vennplot}
\usage{
vennplot(x, names = NULL)
}
\arguments{
\item{x}{list, unique ids for groups}

\item{names}{character, the names for each group}
}
\description{
vennplot
}
