% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_aretes.R
\name{verif_arete}
\alias{verif_arete}
\title{Verification du nombre d'arêtes}
\usage{
verif_arete(g)
}
\arguments{
\item{g}{grille}
}
\value{
si on a le bon nombre d'arrêtes en fonction du chiffre dans la case
}
\description{
Verification du nombre d'arêtes
}
