% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{z.score.vec}
\alias{z.score.vec}
\title{A function to calculate z-scores of a vec}
\usage{
z.score.vec(v)
}
\arguments{
\item{v}{a vector of values}
}
\description{
A function to calculate z-scores of a vec
}
\examples{
vec <- runif(100,2,10)
z.score.vec(vec)
}
\keyword{z-score}
